+++ R/c.r
@@ -42,11 +42,3 @@
 str_c <- function(..., sep = "", collapse = NULL) {
   stri_c(..., sep = sep, collapse = collapse, ignore_null = TRUE)
 }
-
-#' @export
-#' @rdname str_c
-#' @usage NULL
-str_join <- function(..., sep = "", collapse = NULL) {
-  .Deprecated("str_c")
-  stri_c(..., sep = sep, collapse = collapse, ignore_null = TRUE)
-}